backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0